+++ lib/src/platform/file_picker_method_channel.dart
@@ -10,6 +10,7 @@ import 'package:file_picker/src/api/platform_file.dart';
 import 'package:file_picker/src/api/android_saf_options.dart';
 import 'package:file_picker/src/api/windows_options.dart';
 import 'package:file_picker/src/api/linux_options.dart';
+import 'package:file_picker/src/api/web_options.dart';
 import 'package:file_picker/src/platform/file_picker_platform_interface.dart';
 import 'package:file_picker/src/utils/file_picker_utils.dart';
 
@@ -53,6 +54,7 @@ class MethodChannelFilePicker extends FilePickerPlatform {
     AndroidSAFOptions? androidSafOptions,
     WindowsOptions windowsOptions = const WindowsOptions(),
     LinuxOptions linuxOptions = const LinuxOptions(),
+    WebOptions webOptions = const WebOptions(),
   }) => _getPath(
     type,
     allowMultiple,
@@ -81,6 +83,7 @@ class MethodChannelFilePicker extends FilePickerPlatform {
     AndroidSAFOptions? androidSafOptions,
     WindowsOptions windowsOptions = const WindowsOptions(),
     LinuxOptions linuxOptions = const LinuxOptions(),
+    WebOptions webOptions = const WebOptions(),
   }) async {
     try {
       return await methodChannel.invokeMethod('dir', {
@@ -168,6 +171,7 @@ class MethodChannelFilePicker extends FilePickerPlatform {
     bool lockParentWindow = false,
     WindowsOptions windowsOptions = const WindowsOptions(),
     LinuxOptions linuxOptions = const LinuxOptions(),
+    WebOptions webOptions = const WebOptions(),
   }) async {
     try {
       if (onFileLoading != null) {