+++ lib/src/platform/web/file_picker_web.dart
@@ -8,6 +8,7 @@ import 'package:file_picker/src/api/platform_file.dart';
 import 'package:file_picker/src/api/android_saf_options.dart';
 import 'package:file_picker/src/api/windows_options.dart';
 import 'package:file_picker/src/api/linux_options.dart';
+import 'package:file_picker/src/api/web_options.dart';
 import 'package:file_picker/src/platform/file_picker_platform_interface.dart';
 import 'package:file_picker/src/utils/file_picker_utils.dart';
 import 'package:flutter_web_plugins/flutter_web_plugins.dart';
@@ -50,6 +51,7 @@ class FilePickerWeb extends FilePickerPlatform {
     AndroidSAFOptions? androidSafOptions,
     WindowsOptions windowsOptions = const WindowsOptions(),
     LinuxOptions linuxOptions = const LinuxOptions(),
+    WebOptions webOptions = const WebOptions(),
   }) async {
     throw UnimplementedError('getDirectoryPath() has not been implemented.');
   }
@@ -71,6 +73,7 @@ class FilePickerWeb extends FilePickerPlatform {
     AndroidSAFOptions? androidSafOptions,
     WindowsOptions windowsOptions = const WindowsOptions(),
     LinuxOptions linuxOptions = const LinuxOptions(),
+    WebOptions webOptions = const WebOptions(),
   }) async {
     FilePickerUtils.validateAllowedExtensions(type, allowedExtensions);
 
@@ -191,7 +194,7 @@ class FilePickerWeb extends FilePickerPlatform {
     uploadInput.addEventListener('change', changeEventListener.toJS);
     uploadInput.addEventListener('cancel', cancelledEventListener.toJS);
 
-    if (cancelUploadOnWindowBlur) {
+    if (cancelUploadOnWindowBlur && webOptions.cancelUploadOnWindowBlur) {
       // Listen focus event for cancelled
       window.addEventListener('focus', cancelledEventListener.toJS);
     }
@@ -229,6 +232,7 @@ class FilePickerWeb extends FilePickerPlatform {
     bool lockParentWindow = false,
     WindowsOptions windowsOptions = const WindowsOptions(),
     LinuxOptions linuxOptions = const LinuxOptions(),
+    WebOptions webOptions = const WebOptions(),
   }) async {
     if (bytes.isEmpty) {
       throw ArgumentError(